Move test into core/tests
